update scenario model classes so they compile but keep same fields
some have been reduced to stubs, others have had OTP TripPatterns and
TripTimes replaced with equvalent R5 classes, in one case SampleFactory
and Sample were replaced with StreetLayer and Split.

None of these are expected to work as-is, this is just a first step to
ensure that all incoming scenario JSON can be decoded.