Reduce file because speed saved as short
Speed is read as km/h, mph, knots and converted to m/s.
It is multiplied by 100 and rounded (this saves only 2 decimals).

Previously it was saved as int with 7 decimals.

New way of saving lowers file size for 3.5%.
Returned speeds converted to km/h differ only in second decimal.