Small refactoring of permissions
getWayPermissions now returns class with both forward and backward
permissions. Previously it needed to be called twice once for forward
and once for backward. New way is a little faster (since permissions
are read only once from tags) but more importantly it is easier to
support monodirectional permission *:forward *:left *:right

Previous way of calling still exists for now but is deprecated.