Add time to StreetRouter State
Weight and time are currently basically the same.
They differ only on stairs (because of stair reluctance) and when
walking (because of walking reluctance).

In OTP they mostly differ on turn costs and elevation. Both of which
aren't supported yet.

This is very hackish currently since time is read from profileRequest
where it isn't specified if the time is in localTime or UTC. It is
currently assumed to be UTC.

Time is currently saved as Instant in State, to make debugging easier
(since it is printed as date and time instead of epoch in debugger)