Starts work on Transit GraphQL API
On request streetRouter is used to route from start point to all the
stops. Then RaptorWorker is used to route from those stops to all of
the stops(for now)
To found to which stops RaptorWorker needs to route we need to route
from end point to the stops backward.