Add preliminary support for transit response in GraphQL API
Acess, egress and transit segments are created. Segment patterns are
filled with correct departure/arrival times. Itinerary is filled.

Not all the fields are in GraphQl API yet(from, to times mostly)
 and no deduplication happens.

Currently if there are two trips on the same stops two profileOptions
are returned instead of just two different times in same SegmentPattern.