Added preliminary P+R support to routing
Working on #44.
P+R is supported currently only on access. (API needs to be changed
because it can ONLY be possible on access)

Code needs refactoring so that Bike sharing and B+R would also be
possible and path reading will be nicer. Information about P+R also
needs to be added.