Split LegMode in GraphQLAPI for better validation
Since some modes can appear only on acces leg:
- CAR_PARK
- BIKE_PARK
Leg mode was splitted into accesLegMode which is used only on access
legs and otherLegModes which is used on egressLeg and directModes.
With this GraphQL itself does validation and makes sure that CAR_PARK
for example isn't used in egress or directModes.

LegMode with all the modes is still used in returns in streetSegment
since it is outputted from our code.