Changed GraphQL query API
from and to aren't objects anymore but they are split into fromLat,
fromLon toLat and toLon. Requests to server are made with variables in
Demo server now.

TODO: we need to support complex variables. Currently only strings,
floats or booleans can be sent as graphql variables since they can't be
deserialized with Jackson otherwise and I don't think writing full
schema for all the input objects is the right way to do it.