Fares have currencies now and are rounded
Currently fares are set only from last Path that is used. So if there
are two different paths with same stops (so that they are in the same
Profile option) but with different times or different agencies they get
set the same fare. This needs to be fixed.

Working on #69