Add wheelchair permissions when labelling edges
Permission is added if road is walkable unless it is steps.
It is also added if it has wheelchair=yes/limited or
ramp:wheelchair=yes.

Or removed if any of those two are no.

Permission tree is updated becase it made most sense to add wheelchair
as child of FOOT permission.

Tests are also updated.
Working on #109;