Stop search after wanted number of stops/P+R found
Car search in access part and P+R leg is the slowest part of a search.
This speeds it up to stop search after wanted number of stops/P+R stops
are found.

Otherwise Car search found 16k stops which is useless
since only the closest 200 are currently used. This is made with help
of routing visitor which also saves only stops that are farther away
then minTravelTimeSeconds same as used when combining stops.
If stop/P+R is found multiple times state with lowest dominace variable
is kept.

Found stops/vertices are transparently used in
getReachedStops/getReachedVertices(with flags) so that means that on
calling those functions found items are only returned and iteration
over found states isn't needed. This is also transparent for the
calling code.

The only thing needed is to set streetRouter.transitStopSearch to true
or streetRouter.flagSearch to wanted flag.
It is also possible to change how many stops should be found (default
is value of PointToPointQuery MAX_STOPS) and how many vertices with
flags (P+R/bike shares) which is 20 currently.
This was also done in this commit.

TODO: use this also in Bike shares