Add spatial index to pointsets. Use it when building distance tables.
Time complexity for distance table computation is now O(nStops) instead of O(nStops * nPoints), i.e. the run time will no longer depend on the size of the pointset. This should change distance table computation from one of the slowest parts of scenario application to one of the fastest parts. We used to iterate over all points for each stop, then added filtering of those points during the iteration based on a bounding box, which did not improve speed. A spatial index makes for a massive improvement though.