feat(analysis): serialize grid pointset and its linkage in analysis networks
Addresses #217.
The GridPointSet and its linkage are no longer lazy-initialized by getter functions.
In networks intended for analysis work (created by the TransportNetworkCache) the GridPointSet and its linkage are pre-built before saving the network.
They both stored in non-transient fields (rather than Guava caches) to ensure they are serialized.
They are updated when a scenario is applied, so the new scenario network has a matching GridPointSet and linkage.
Marked StopToPointDistanceTables non-transient so they are also stored when the network is saved.
This all means we’ve got quite a different, fast-startup mechanism in use for the newer grid-based analysis, but older freeform PointSet analysis should still work at a slower pace.