Add support for showing turn restrictions in map
Debug map has a new layer turns. Where turn restrictions are shown.
Currently they are shown with same colors as permissions. In addition
to permission layer properties they have edge property which can be
(FROM, VIA, TO) and only property which is true if it is ONLY TURN
restriction or false otherwise. (Meaning it is no turn restriction)