Turn restriction are correctly updated in scenario
If adding stops in scenario splits from/to/via edges Turn restrictions
need to be updated in new transportNetwor. But they need to stay the
same in original. For this we copy all 4 turn restriction fields. And
make copy of each TurnRestriction which needs new from/to/via edges in
this copied list. We could copy on write since chance that turn
restrictions needs to be copied is small, but Turn Restrictions are
small and if need arises that COW is needed we would do it.

Tests are also created