fix projection on GeoTIFF export
There was an assumption in the code that edges of a Mercator rectangle were parallel to edges of a WGS84 rectangle.
This was a bit of a challenge to work out because web Mercator coordinates are often in pixels, but when you want to store them independent of zoom level they're often in meters, and different sources claim different world widths / heights in meters.