fix(grid): Add 1 to grid height and width to prevent data loss and /0
Grids could have a height of zero if all of the data was at the same latitude which would cause an
empty grid. Also, any data at the bounds of the grid would be left out since it would be equal to
the height or the width where our bounds check requires it to be less than that.