short polling on broker
this is a major change. instead of holding connections open, workers short-poll and sleep if they receive no work.
the way results are reported also changes. The new methods in the broker are wrapped in HTTP API methods in the analysis backend.
this change also stops writing results to S3, sending them straight back to the backend.
We'll still want to save regional results to S3 (or somewhere) but we can probably even pass back results for a single origin in the worker reporting JSON