remove bootstrapping from the accessibility reducer, fix minCount computation
Addresses #327, #328.
The definition of percentiles was changed to match that in the travel time grid reducer.
It is now based on the distance between the first and last value (n-1) rather than the total number of values (n).
This fits our expectations of the 0th and 100th percentile as min and max.
Note that the change in percentile definition *increases accessibility results*, in some places significantly.