return reducer results up the stack
We were passing lots of things like output streams down a chain of function calls and writing out the results as a side effect of computation calls.
This has been changed so that the two reducers return the same compound type, which is then also returned by the propagater and the TravelTimeComputer.
The writing is then an action taken on these results.
This moves us closer to having a single reducer type, then no reducer at all (when it's merged into the propagator) and being able to produce both kinds of results at once on a single task.