Restructure static site regional metadata
This splits the metadata into two files, transitive.json and request.json. It also removes the zeros that were mistakenly prepended to the metadata files.
Functionality for serializing objects as JSON into a persistence buffer are factored out.
The extents of the static site used to be (incorrectly) replicated in the top level metadata object and in the request it contained.
The top level object recorded the extents of the entire street network, while the embedded request object had the extents actually used for the analysis.