feat(street-router): set time and distance for initial edge
The actual feature was implemented in 0c9e19a ("fix(street-router)...").  This is just a dummy comment change to trigger a new miner version number.

AFFECTS RESULTS: Adds the distance from an origin to the nearest street intersection, which was previously ignored.  This may reduce the area reachable and number of opportunities accessible within a given time cutoff, especially in areas with large blocks/long distances between intersections.