fix(async-loading): remove flawed synchronous methods
Instead of having both single point and regional tasks go through the pre-loader, the regional tasks just bypass it entirely now and hit the network cache directly.
The problem with the synchronous preloading code was that it triggered the preload process on every request, setting partial load progress even when loading had already been previously accomplished.
It then never hit the point where the loading state was PRESENT, which premanently prevented any subsequent async requests from succeeding.
The ideal was to have both kinds of tasks go through the preloader so we could eventually get rid of the old nested loadingCaches, but that's going to take more refactoring.