fix(add-trips): set route type #465
The mode of travel for added trips (routes) still can't be set in the UI, but this change at least copies over the default value (bus) from the AddTrips modification class.
This has been confirmed to work - when bus is deselected in the UI, added routes are not considered.
It should suffice to include a mode property in the modification JSON.