fix(TransportNetworkCache): use lenient object mapper when reading scenarios
Workers use a lenient object mapper that ignores unrecognized fields when
deserializing scenarios embedded in single point requests. We want to apply
the exact same deserializer for scenarios read from disk and S3, for regional
analyses.

The immediate reason for this is that the branch supporting custom modifications
no longer has a type field in Modifications. The ModificationTypeResolver reads
the type field, but there is no target field for it to be deserialized into.

We may want to also add an annotation to ignore this field during deserialization.