fix(TravelTimeComputer): non-transit uses same flow as transit access
In searches with no transit modes, the origin point was not being set.
I have fixed this by just using the same code path for both transit
and non-transit searches. This change also means that the maximum travel
time for direct on-street travel is the same as for access to transit.
This is more coherent.

Also includes a lot of documentation updates.
Also (importantly) reinstates park+ride support, though still in a way
that excludes use of any other modes.