fix(shapefile): extend shapes before performing noding
Previously, it was noding before extending roads, because the intent was
to detect and extend only dead-end segments after a first pass of noding.
No reasonable way was found to detect post-noding dead ends, so instead
we're extending every shape by a few meters before noding. This makes
the network significantly more complex but it works.