fix(points): write travel times when requested
We want R5 to make isochrones (create a TimeGrid) for single-point and taui tasks.

We want R5 to create a TravelTimeResult when travel times between freeform pointsets are requested.

In either case, travelTimes will be set to a non-null value, and we want to set travel times at the targets.