refactor(RegionalTask): replace nTravelTimeTargetsPerOrigin with method
Besides being derived data (which can be calculated from other fields on
demand) it was only used some of the time. It's getting difficult
to keep track of when certain fields are meaningful or not, so we should
progressively eliminate them or at least provide full javadoc on when
they are active.