fix(ShapefileReader): unlock shapefile properly by closing iterator
Note this doesn't work on partial iteration (e.g. due to an exception).
But it's better than leaving it locked every single time we open it.
Based on past experience, the lock doesn't impede execution but causes
loud error log entries. Also added a comment on the weirdness that led
me to look at shapefile locking.