fix: derive mapdb file name from osm file name
During tests (or even during regular use) we could have
more than one OSM file in the same directory. if you try
to load more than one file at once in the same
directory they would all try to write to the same file.

This became noticeable when tests were run in parallel
(via forkCount in the POM).

Also lowered that forkCount in case our testing
environment doesn't have a lot of memory, and added a
comment to the POM.