fix(egress): consolidate to single short-circuit
Egress areas in a pickup delay modification can restrict the availability of a given egress mode. But they do not have an associated rebuildZone, so egress cost tables for the baseline (rather than scenario) were being returned.

Rather than customizing the rebuildZone logic, this commit disables the return of base linkage short-circuit if a pickupDelay modification is present. It also consolidates the check for availability of egress modes (and removes a method that is unused after this consolidation)