basic error diffusion when writing grids #566
Diffusion has directional bias (eastward) and up to one half opportunity
can be lost per row of pixels. This is a best-effort approach
maintaining backward compatibility. It is probably not worth improving
with iterative diffusion, we would instead migrate to a file format
supporting fractional counts eliminating the need for diffusion.