Move freeform PointSet lookup out of synchronized method
Freeform pointset use is becoming mainstream, we don't want to be
stalling the broker on production by doing potentially slow things in
constructors called in the broker's synchronized methods.

We make sure any PointSet instances we need are preloaded into the
transient fields of the template task on the backend, where they are
visible to the Job and the MultiOriginAssembler.