expose regional errors to UI and cancel job #517
Includes backend (result assembler) errors, as well as errors that
occur on the workers.
Adds error fields to RegionalWorkResult, Job, and JobStatus.
Errors are deduplicated and stored on the Job, which is then considered
inactive, ending the delivery of tasks to workers, allowing them to shut
down.
Some cleanup of Broker methods for clearer synchronization.
Also deleted unused RegionalAnalysisStatus class to avoid confusion.