factor out two boarding / reboarding search methods
One is used for the special case of reboarding with no overtaking.
The other is used for first boarding and reboarding with overtaking.

This avoids the potentially confusing reuse of variables identifying the
currently boarded trip to tentatively hold replacement trips within
reboarding searches.

Symbolic constants are used everywhere for "no trip" or "no time".