enforce read-only GTFSFeeds via factory methods
This will fail fast if we write to feeds returned from GtfsCache.
We no longer (re)create patterns when building TransitLayers, the input
feeds are treated as immutable. We still leak some writable feed objects
outside gtfs-lib for tests and progress reporting, further refactoring
would be necessary to avoid this.