key random offsets on TripSchedules (issue #740)
this is a minimal fix that builds a secondary map, avoiding translation
from filtered to unfiltered pattern and trip indexes.
There are better ways to do this but they all seem to cascade series of
refactors so carry some risk of introducing more bugs.
This one is easier to test and trace, so makes a good first step.