handle single point worker shutdown special case
This was observed in testing when running many single point and regional
analyses. Perhaps a regional worker started up and was assigned the
same IP address as a single point worker that had recently shut down
from inactivity. The resulting exception must be caught so this IP will
be unregistered for this workerCategory and a new one started/found.