cleanly handle non-existing regional result grids
Ideally we'd infer from the database record that this analysis does not
have any gridded results and immediately return a 404. As a short-term
fix, we check assertions only when an older-format file is actually
found, and return a 404 to cover a wider range of cases: files that are
missing because they are not even supposed to exist, and files that
should exist but have gone missing for other unknown reasons.