Revert "Revert "Use getPriority method instead of a field""
This reverts commit f399f90ee496b4498d8d55f49f6d7966c02b46af.
This is a double-revert, re-establishing the original use of
getPriority methods instead of priority fields. The hastily
applied patch where we shifted to using the field instead of
methods was based on a false belief that old MapDBs had this
field. In fact the field was a temporary addition during
development, and the original state was without the field.