do not start HTTP server in constructor
This is a first step in refactoring the worker.
It pulls out a Worker class that can be used by both the standalone
main method used by cluster workers and the thread worker in local use.
It's only initializing two components now, but provides a proper place
to share code initializing worker components in the right order.
It's also a first step toward clean separation of concerns between a
regional work processor and the worker's synchronous HTTP API.