Factor out common vector tile code
This change factors out common code across the `GtfsVectorTileMaker` and the `NetworkTileController`. It moves manny common functions into the `MapTile` util. This is because many operations are dependent on the `envelope` that is produced by that util. Additionally, the remaining functionality of the `GtfsVectorTileMaker` was eseentially boiled down to just creating and querying the caches.