Remove MongoJack and consolidate utils
A process that started many years ago, this removes the MongoJack dependency and migrates the `Persistence` module and its collections to the `AnalysisDB` and `AnalysisCollection` types that were set to replace it.

In light of a possible database switch in the near future, it seemed more straightforward to continue to use `String` based IDs everywhere, instead of switching to a MongoDB specific `ObjectID` type.

Our average collection already uses `String` IDs, but this will require a migration for aggregation areas, data sources, and data groups.