Convert FileItems to Files before using
We use Apache FileUpload to assist with handling incoming files. We also use a DiskItemFactory that persists each incoming file to disk immediately. This means that we could represent each incoming file as a java.io.File instead of a FileItem. java.io.File is a much more common way of handling a file and easier to work with.