fix final and static field modifiers in broker
Many constants (as indicated by capitalization) were not static.
One field always set in the constructor was not final.
The list of resultAssemblers WAS static. This may be a remnant of pre-
component design. This was not problematic because we never construct
more than one Broker, but it should be possible to do (no global state).