Exclude modifications for regional analysis lookup
This change adds a new helper method to pass a `fields` object to `MongoMap#findByIdIfPermitted` and utilizes that to exclude modifications while looking up a regional analysis to generate a `scenarioJsonUrl`.

This will prevent errors from occurring if custom modifications are used in a worker version that do not exist in the server.