Validate `dualAccessThresholds` before using
When we generate results with temporal densities (either CSV or grid) we now require `dualAccessThresholds` to be set. This change adds a validation method (similar to cutoffs and percentiles) and calls that method when creating a regional analysis.

Additionally, this moves the `SemVer` check to only allow r5 > v7.4 when using `includeTemporalDensity`. Otherwise, the old `dualAccessiblityThreshold` values required by old worker versions will be silently ignored.